Move generator to vosk.py
